Implement better accessibility. It's far from perfect; see extended description. Issue #8
Greek words don't pronounce correctly, and I'm not sure there is any
way to make them (at least not for polytonic greek - I tried el-polyton
with no good result.)
If VoiceOver is off and then turned on, the elements are not calculated
until the labels are added to the cell again. For now, VO should be on
prior to starting the app.
Tapping and missing all elements results in a seemingly random element
selected.
Tap areas may be too small.